moved softdash into site to start to "fix" it
I moved the softdash artefacts for the style and structure of the site
to where I believe they should go.  This will produce lots of errors and
I will now start to fix them as I find them through the automatic
building of the site.

Already have Sass errors as the files are in the wrong directory.

I will start here first.