menu changed through static code
There is a data file for this but it doesn't work.

I am noticing a lot of broken parts and although this theme demos well,
it is a bit of a mess.  A lot of work will be needed to get it up to
scratch, but there aren't a great many visitors so I shall do that once
gone live.

For live I will present a front page to display blogs and an about page.

I will style them to look better later