nsferred over to Mundana them for snagging
The next step is to fix up the posts and pages and also understand what
I can do with this theme.

I will not be publishing it on the internet until there is enough
content to make it worth while.  As of yet I don't know how much enough
is.