removed theme menu items from main navigation
The makers of the theme had links totheir sites and this removed them.

It has left the `Home`, `Authors` and `Contact` links.

The `Contacts` and `Authors` needs some thought as to whether I want or
need them.  In fact I need to decide what links go up there.  An About
with Privacy Policy and GDPR would be useful.