in preparation for short post series
recently I have considered removing octopress from this and all my
web sites on the grounds it hasn't been updated in agaes and is unlikely
to be updated again.

I started this process here and elsewhere, but having had a look at the
octopress gem, I'm not sure it needs updating unless jekyll and\or Ruby
change.

As a result I'm going back to octopress and setting my sites up for it
to use it the way I want it to.

That's the first part of this commit.

I know, I know one thing at a time but this is life.

The other part is renaming my unpublished post extension to use .md
rather than .markdown.  This is to bring it inline to how I want
octopress to create the file extensions for posts, pages and drafts.

I'm also going to change the potentially large post about using REST
with BI Publisher to at least two different posts.